+++ changelog.md
@@ -4,6 +4,7 @@ dev
 ------------------
 - select_from_text_positions_without_nesting doesn't accept RangeSet
 - Enactment.limit_selection.start must be an int
+- No separate LinkedEnactment class for Enactments with URL links as children
 
 0.5.2 (2021-05-20)
 ------------------